refactor(project): rename Project to IProject across domain, component, and pages
